release-kit|refactor: Extract readCurrentVersion and simplify parseCanonicalSemver
Extracts the `readCurrentVersion` helper (and its `hasVersionField` type guard) from `releasePrepare.ts` and `releasePrepareMono.ts` into a shared internal module `readCurrentVersion.ts`, removing duplicated logic between the single-package and monorepo flows. Simplifies `parseCanonicalSemver` in `compareVersions.ts` by dropping the redundant second undefined-check after a successful regex match — the `CANONICAL_SEMVER_PATTERN` has exactly three capture groups, so a successful `.match()` already guarantees the groups are present. These changes are behavior-preserving polish.
