+++ packages/release-kit/src/releasePrepare.ts
@@ -1,5 +1,4 @@
 import { execSync } from 'node:child_process';
-import { readFileSync } from 'node:fs';
 
 import { bumpAllVersions, setAllVersions } from './bumpAllVersions.ts';
 import { isForwardVersion } from './compareVersions.ts';
@@ -9,6 +8,7 @@ import { generateChangelogJson } from './generateChangelogJson.ts';
 import { generateChangelogs } from './generateChangelogs.ts';
 import { getCommitsSinceTarget } from './getCommitsSinceTarget.ts';
 import { hasPrettierConfig } from './hasPrettierConfig.ts';
+import { readCurrentVersion } from './readCurrentVersion.ts';
 import type { BumpResult, Commit, PrepareResult, ReleaseConfig, ReleaseType } from './types.ts';
 
 /** Options for the release preparation workflow. */
@@ -161,22 +161,4 @@ export function releasePrepare(config: ReleaseConfig, options: ReleasePrepareOpt
     formatCommand,
     dryRun,
   };
-}
-
-function hasVersionField(value: unknown): value is { version: string } {
-  return typeof value === 'object' && value !== null && 'version' in value && typeof value.version === 'string';
-}
-
-/** Read the `version` field from a package.json file. Returns undefined if the file can't be read or parsed. */
-function readCurrentVersion(filePath: string): string | undefined {
-  try {
-    const content = readFileSync(filePath, 'utf8');
-    const parsed: unknown = JSON.parse(content);
-    if (hasVersionField(parsed)) {
-      return parsed.version;
-    }
-  } catch {
-    // Return undefined if the file can't be read or parsed.
-  }
-  return undefined;
 }
